vga/bg_rast: avoid race in PendSV
Turns out, the write to the NEXT_XFER fields fell late enough in hblank
that they can overlap the SAV IRQ. When this occurs, the SAV ISR
attempts to access the fields, and panics.

It also turns out that the order of operations in this part of the
PendSV ISR is not significant -- I can simply move the racing operation
earlier. Now there's a small chance for clash between the end of the big
copy_words transfer and SAV, but such a clash is much less scary than
potentially reading the wrong next_dma_xfer value (for example).

**This race appears in the C++ code** and may explain some intermittent
issues I've seen in mode changes.