util/rw_lock: improve preemption behavior
This is now much harder to use incorrectly. The motivation here is that
the exclusive monitor -- the processor state used by LDREX/STREX -- is
cleared by interrupt entry on Cortex-M. So the STREX can fail if *any*
interrupt occurs, not just one that uses the lock.

That's a potential source of intermittent flake.

The code generated by this implementation is a bit worse, but it seems a
small price to pay for correctness. (And it's not actually awful.)