tunnel: fix numerics, optimize
The C++ demo actually relies on undefined right-shift behavior in C++
for the shading algorithm. Trying to do that in Rust got *different*
undefined behavior, causing glitches in the shadowy parts of the center.
I've rewritten it to stay entirely within the bounds of defined
behavior, which has made it about 12% slower after optimization.