rook: more aggressive quantization -> fewer segments
It turns out that points in the rook only need one decimal digit on the
right of the decimal point to look smooth and convincing. This collapses
a bunch of fine segments together, reducing segments drawn by 8% and
vertices transformed by 18%.

Unfortunately, the removed segments tended to be very short -- that's
why they were removed -- and so this does *not* improve performance by
an even 8%.

I also noticed that the brag line was understating the number of
triangles by 1000, a mistake I copied from C++. Fixed.